Revamp actionpack-xml_parser gem
Since Rails 5, parameter parsing is not done in a middleware. This gem
was using the middleware to do the parameter parsing so it was failing.

This gem now only works with Rails 5 so we can use the new API to
register a new parameter parser.

It was removed the support to legacy HTTP_X_POST_DATA_FORMAT to set the
content type.